% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liss_credentials.R
\name{liss_delete_credentials}
\alias{liss_delete_credentials}
\title{delete stored LISS Data Archive credentials}
\usage{
liss_delete_credentials(username)
}
\arguments{
\item{username}{character or numeric. the username whose credentials
should be removed.}
}
\value{
invisible \code{TRUE} on success.
}
\description{
removes the username and password previously saved with
\code{\link[=liss_store_credentials]{liss_store_credentials()}} from the system keyring.
}
\examples{
\dontrun{
liss_delete_credentials(username = 12345)
}
}
