% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liss_credentials.R
\name{liss_list_credentials}
\alias{liss_list_credentials}
\title{list stored LISS Data Archive credentials}
\usage{
liss_list_credentials()
}
\value{
a data frame with columns \code{service} and \code{username}.
}
\description{
returns a data frame of usernames with stored credentials.
}
\examples{
\dontrun{
liss_list_credentials()
}
}
