% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liss_server_status.R
\name{liss_server_status}
\alias{liss_server_status}
\title{check LISS Data Archive availability}
\usage{
liss_server_status(verbose = TRUE)
}
\arguments{
\item{verbose}{logical. print status messages to the console?}
}
\value{
invisible logical. \code{TRUE} if the website is accessible.
}
\description{
sends a HEAD request to the LISS Data Archive website and reports
whether the server responds with HTTP 200.
}
\examples{
\dontrun{
liss_server_status()
}
}
